SIG: Move xnet to internal
