+++ go/sig/internal/xnet/BUILD.bazel
@@ -3,7 +3,7 @@ load("@io_bazel_rules_go//go:def.bzl", "go_library")
 go_library(
     name = "go_default_library",
     srcs = ["xnet.go"],
-    importpath = "github.com/scionproto/scion/go/sig/xnet",
+    importpath = "github.com/scionproto/scion/go/sig/internal/xnet",
     visibility = ["//visibility:public"],
     deps = [
         "//go/lib/common:go_default_library",
