+++ go/sig/internal/xnet/xnet.go
